feat(homepage): overhaul layout and introduce new sections for enhanced user experience. refactoring!
- Updated the layout title and description for clarity and branding consistency.
- Refactored the HomePage component to utilize new sections: Hero, Features, Pricing, Chat Demo, Analytics, CTA, and Footer.
- Introduced modals for email signup and mascot vibes, improving user engagement.
- Added animated background for a dynamic visual experience.
- Integrated hooks for managing homepage state and content data, enhancing maintainability and responsiveness.
- Created new components for event cards, feature cards, pricing cards, and statistics, streamlining the UI.
