+++ src/components/ui/AnimatedBackground.tsx
@@ -1,0 +1,14 @@
+import type React from "react";
+
+export function AnimatedBackground() {
+  return (
+    <div
+      id="background-elements"
+      className="fixed inset-0 overflow-hidden pointer-events-none"
+    >
+      <div className="absolute top-1/4 left-1/4 w-64 h-64 bg-brand-primary/10 rounded-full blur-3xl animate-pulse"></div>
+      <div className="absolute top-3/4 right-1/4 w-96 h-96 bg-brand-secondary/10 rounded-full blur-3xl animate-pulse delay-1000"></div>
+      <div className="absolute top-1/2 left-1/2 w-80 h-80 bg-blue-500/10 rounded-full blur-3xl animate-pulse delay-500"></div>
+    </div>
+  );
+}
